+++ core/slru.go
@@ -13,12 +13,16 @@ type slruBase[K comparable, V any] struct {
 	// protected is the warm region of slru cache.
 	protected *lruBase[K, V]
 
+	// promotion represent the type of promotion from
+	// probationary to protected.
+	promotion PromotionType
+
 	// stats represents the metrics added by its private methods
 	stats Stats
 }
 
 // assembleSLRU creates an instance of [slruBase] using the given capacity and ratio.
-func assembleSLRU[K comparable, V any](capacity int, ratio int) (*slruBase[K, V], error) {
+func assembleSLRU[K comparable, V any](capacity int, ratio int, promotion PromotionType) (*slruBase[K, V], error) {
 	if ratio < 0 || ratio > 100 {
 		return nil, ErrInvalidSLRURatio
 	}
@@ -37,12 +41,22 @@ func assembleSLRU[K comparable, V any](capacity int, ratio int) (*slruBase[K, V]
 	return &slruBase[K, V]{
 		probationary: prob,
 		protected:    prot,
+		promotion: promotion,
 	}, nil
 }
 
 // Capacity returns the maximum allocated capacity of the SLRU cache.
 func (s *slruBase[K, V]) Capacity() int {
 	return s.probationary.capacity + s.protected.capacity
+}
+
+// Close safely terminates the instance and frees up the memory.
+func (s *slruBase[K, V]) Close() {
+	s.protected.clearState()
+	s.protected.stats = Stats{}
+
+	s.probationary.clearState()
+	s.probationary.stats = Stats{}
 }
 
 // Contains checks whether the key is present in the Cache.
@@ -244,12 +258,8 @@ func (s *slruBase[K, V]) getWithKey(key K) (V, bool) {
 		}
 		s.probationary.addHits()
 
-		val = s.probationary.deleteWithIndex(curr).value   // this won't trigger an eviction
-		if len(s.protected.hash) == s.protected.capacity { // demoted key
-			kv := s.protected.removeOldest() // this will trigger an eviction in protected
-			s.probationary.addKey(kv.key, kv.value)
-		}
-		s.protected.addKey(key, val) // promoted key
+		val = s.probationary.deleteWithIndex(curr).value // this won't trigger an eviction
+		s.promoteKey(key, val)
 	}
 	return val, true
 }
@@ -286,18 +296,34 @@ func (s *slruBase[K, V]) peekWithKey(key K) (V, bool) {
 	return s.protected.peekWithIndex(curr).value, true
 }
 
+// promoteKey promotes a key from probationary to protected and demotes
+// a key if necessary. It ensures that the probationary key is already deleted
+// to have space for the protected key.
+func (s *slruBase[K, V]) promoteKey(key K, value V) {
+	if len(s.protected.hash) == s.protected.capacity { // demoted key
+		kv := s.protected.removeOldest() // this will trigger an eviction in protected
+		s.probationary.addKey(kv.key, kv.value)
+	}
+	s.protected.addKey(key, value) // promoted key
+}
+
 // putWithKey adds or updates the cache value with key "key".
 // It also returns a value based on whether it was evicted or replaced.
 func (s *slruBase[K, V]) putWithKey(key K, value V) (UpsertState, V) {
 	curr, ok := s.protected.retrieveIndexWithKey(key)
 	if !ok {
-		return s.probationary.putWithKey(key, value)
+		if s.promotion == PromotionGet {
+			return s.probationary.putWithKey(key, value)
+		}
+		curr, exists := s.probationary.retrieveIndexWithKey(key)
+		if !exists {
+			return s.probationary.putNewKey(key, value)
+		}
+		val := s.probationary.deleteWithIndex(curr).value // this won't trigger an eviction
+		s.promoteKey(key, value)
+		return UpsertReplace, val
 	}
-	val := s.protected.peekWithIndex(curr).value
-
-	s.protected.updateWithIndex(curr, value)
-	s.protected.makeRecent(curr)
-	return UpsertReplace, val
+	return s.protected.putOldKey(curr, value)
 }
 
 // removeOffsetIndex, offsets the index for protected.